c11 _Generic __typeinfo to const char * str test
An experiment showing that C11's _Generic can be used, albeit awkwardly, to get
type info as strings, then generate an argument-type static array to pass to a
function along with the macro varargs. The called function can then consume the
arguments with va_ functions or pass them to vsprintf etc.